Remove "Project Administrators"
This distinction had been a remnant of the past back when we still had a
group of developers and principal developers. When we made everyone a
principal developer we renamed the latter group to "project
administrators".

As such I feel that we should abandon this arbitrary split and simply
remove the title "project administrators". Mostly because I feel it
somewhat defeats the purpose of making everyone an equal principal
developer in the first place.